requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
torch
torchvision
transformers
segment-anything
supervision
tqdm
matplotlib
streamlit-image-coordinates
watchdog
opencv-python-headless==4.7.0.72
pyopengl
pythreejs
